Add supplier seeding functionality to DataSeeder class
This commit introduces a new method for seeding initial suppliers into the database within the DataSeeder class. It includes the addition of the SupplierRepository and creates a diverse set of suppliers with realistic business information. Logging statements are added to track the seeding process and confirm successful completion of the operation.